Add zig build and integrate with CI
This commit adds zig build and mimics cmake where it makes sense. There
are some differences. As an example zig build does not support
installing the build onto the system. It also will always include the c
headers as the main purpose of this is to allow c/c++/zig projects to
use the zig package manager to interface with enkiTS.